Add basic image carousel package
This package contains functions to streamline setting up an image
carousel. The factory takes in an element that represents an 'image
slider', or basically a container for multiple images aligned
horizontally.

Users may also add custom functions that execute before and after a
timer automatically swaps to the next image (if the user provides a
timeout duration).